Variable Explorer: Move objecteditor to the varexp namespace
