+++ spyderlib/widgets/varexp/objecteditor.py
